move xhr to utils
