+++ src/components/index.js
@@ -28,4 +28,3 @@ import './Tabs'
 import './TextOverflow'
 import './Tooltip'
 import './Upload'
-import './xhr'
